Moved counters to SignalCounters
